% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_demo.R
\name{generate_demo}
\alias{generate_demo}
\title{generates a demo project built using the archetyper library}
\usage{
generate_demo()
}
\description{
generates a demo project built using the archetyper library
}
\examples{
generate_demo()
}
