% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_demo.R
\name{generate_demo}
\alias{generate_demo}
\title{generates a demo project using archetyper}
\usage{
generate_demo(path = ".")
}
\arguments{
\item{path}{The path where the project should be created. Default is the current working directory.}
}
\description{
generates a demo project using archetyper
}
\examples{
generate_demo()
}
